Add Hero2 front panel LCD support
Enable SPI in the kernel, configure SPI GPIOs, and add a
userspace program to display a logo on the front-panel LCD
on the Hero2 camera. LCD commands are transmitted in 9-bit
mode using the 'spidev' user-mode SPI interface on
/dev/spidev0.0. Refer to the ST7885 datasheet for details.

Update prebuilt kernel, initrd, kernel patches, and kernel
configuration file accordingly.

Signed-off-by: evilwombat <evilwombat@server.fake>