gpboot: Add Hero4 DDR init sequence
Add the command sequence needed to initialize the memory
controller on the Hero4 Silver (and presumably, Black).
Although the RTOS seems to live at 0x03300000, the DDR is
initialized at 0xC0000000, as on earlier cameras. Perhaps
the sequence needs to be modified, or there is something
else going on here. Either way, we don't have anything
useful to actually put in DDR yet, but it's good to have
the sequence working anyway, in case future development
calls for it.

Signed-off-by: evilwombat <evilwombat@server.fake>