gpboot: Switch to Hero3 v300 BLD for HW initialization
When booting Linux on Hero3 Black or Hero3+ Black, we used
to rely on the Hero2 BLD to partially initialize the
hardware, so that our kernel would be happy. This is not
only annoying (requiring deprecated firmware) but may also
result in incomplete or suboptimal hardware initialization.

Switch the Linux booting code to use the Hero3 v300 BLD for
initializing the hardware, to hopefully keep the hardware
happier. Using the Hero3 v300 BLD requires that we poke at
(what appears to be) the I2S MUX register, to keep FIO DMA
requests in the kernel from failing. At least, that's what
it seems like, as we lack any real hardware documentation
for this stuff...

Signed-off-by: evilwombat <evilwombat@server.fake>