Linux: Add A7 kernel with NAND partitioning disabled
Add a version of zImage-a7 which has CONFIG_MTD_PARTITIONS
turned off. This can be used to access the complete NAND
device on a Hero3 Black camera if the NAND partition table
(PTB) is corrupted.

Currently there is no way to specify which kernel to load,
so you would manually need to rename this file to
'zImage-a7' to load it.

Signed-off-by: evilwombat <evilwombat@server.fake>