evilcortex: Add Hero4 (Cortex A9) CPU bootstrap code
When booting the Hero4, we load code over USB using the
ARM11 CPU. However, Linux will actually be running on the
Cortex A9 CPU(s). After we are done loading our software
into DDR, we use the ARM11 to bring the Cortex A9s out of
reset, so it (they) can execute our newly-loaded code.

However, the Cortex A9 CPU(s) power up in a somewhat
indeterminate state, and we need to execute some bootstrap
code on it. This code needs to:
 * Invalidate instruction and data caches, which power up
   in an unknown state
 * Invalidate and initialize the PL310 L2 cache controller,
   which is not part of the CPU itself
 * Enable the SMP bit in ACTLR, to allow coherent memory
   access between the two A9 processors
 * Determine whether we are running on Core0 or Core1, and
   do the right thing. Core0 needs to set up r0-r2 to pass
   the appropriate ATAGS pointer and machine ID to Linux,
   whereas Core1 needs to jump to a special bootstrap loop
   and busy-wait until Core0 tells it that Linux has booted
   far enough to be able to accept additional CPUs.

Signed-off-by: evilwombat <evilwombat@server.fake>