gpboot: Add option for raw-booting a Hero4 image
Add an option for loading and booting a raw image on Hero4
cameras. This is useful for developing and testing code
that would normally run directly on the metal. The raw
image is loaded at address 0x03300000 and is executed by
the Cortex A9 CPUs, while the ARM11 busy-waits forever.

Signed-off-by: evilwombat <evilwombat@server.fake>