gpboot: Add Hero4 recovery feature
Add an option to load a Hero4 "recovery" kernel/ramdisk,
which will automatically attempt to reflash the firmware on
a Hero4 camera. The firmware to be reflashed will likely be
read from the sdcard, as it is too big to load over USB
within a reasonable amount of time. For now, we will be
using the same initrd for standard Hero4 Linux and for the
recovery mode, with the auto-reflash feature being
controlled by a kernel commandline option, which is read
by the userspace startup script.

Signed-off-by: evilwombat <evilwombat@server.fake>