hero4: Add recovery payload file
When performing H4 recovery, the stock firmware is too big
to load over USB within a reasonable amount of time. So,
add a pre-packed payload file that the user can copy onto
their sdcard (and leave in the camera) for use during the
recovery operation. This payload contains an unpacked
version of the Hero4 v3.00 firmware files, a compatible PTB
image, and a shell script that automatically erases and
reflashes NAND (up to but not including the Calibration
partition).

Signed-off-by: evilwombat <evilwombat@server.fake>