gpboot: Support booting Linux / recovery on the Hero5 Black
Add basic Hero5 Linux support. This involves loading a Linux
kernel, a ramdisk, a device tree, and a small hardware init
stub which sets up the machine type and DTB pointer prior to
jumping to the kernel.

Two boot options are supported: booting Linux in standard
mode, and booting Linux in recovery mode. The only difference
between them is the choice of DTB file to pass to the kernel.
The only difference between the DTB files is an 'autoreflash'
flag that is passed to the kernel in recovery mode, to tell
userspace to look for a recovery payload on the memory card.

Signed-off-by: evilwombat <evilwombat@server.fake>